Refactored so input system is more maintainable
