+++ client/Assets/Input/Mobile/MobileControlRig.cs
@@ -10,11 +10,6 @@ namespace UnityStandardAssets.CrossPlatformInput
     [ExecuteInEditMode]
     public class MobileControlRig : MonoBehaviour
     {
-        // this script enables or disables the child objects of a control rig
-        // depending on whether the USE_MOBILE_INPUT define is declared.
-
-        // This define is set or unset by a menu item that is included with
-        // the Cross Platform Input package.
 
 #if !UNITY_EDITOR
 	void OnEnable()
