+++ client/Assets/Input/VirtualAxis.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 54b1b013b97d7bd48aa2c2d83f941c92
-timeCreated: 1467552428
+guid: 1f3c58f7bde7ced4d95d39aa048aef5e
+timeCreated: 1467552084
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
