+++ client/Assets/Input/VirtualButton.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 54b1b013b97d7bd48aa2c2d83f941c92
-timeCreated: 1467552428
+guid: 4f0e199fea59777428971e135cc5fcee
+timeCreated: 1467552084
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
