+++ client/Assets/Sprites/GUI/Mobile/Controls/ButtonThumbstickOverSprite.png.meta
