+++ client/Assets/Sprites/GUI/Mobile/Controls/ButtonThumbstickUpSprite.png.meta
